requirements.txt
google-adk>=0.1.0
google-generativeai>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.12.0
